move and rename service objects
